Array Implementation + get() returns data type now
After creating the DataType enum it's made it a breeze to implement array support as a writeable entry for files. Therefore, I've updated these classes to do that very thing.

It is also now possible to directly retrieve a specific data type when using DataEditor.get(). You simply need to pass an additional DataType argument. It must be the same data type as what's in the file otherwise it will return an error.